Split up interpreter module into smaller grouped modules
