+++ serde_arrow/src/internal/serialization/interpreter.rs
@@ -1,3 +1,6 @@
+mod misc;
+mod primitives;
+mod sequences;
 mod structures;
 
 use std::collections::HashMap;
@@ -5,23 +8,11 @@ use std::collections::HashMap;
 use crate::{
     base::Event,
     internal::{
-        common::{
-            DictionaryIndex, DictionaryValue, MutableBitBuffer, MutableCountBuffer,
-            MutableOffsetBuffer,
-        },
-        conversions::{ToBytes, WrappedF16, WrappedF32, WrappedF64},
+        common::{MutableBitBuffer, MutableCountBuffer, MutableOffsetBuffer},
         error::{fail, Result},
         serialization::{
             bit_set::BitSet,
-            bytecode::{
-                dispatch_bytecode, Bytecode, LargeListEnd, LargeListItem, LargeListStart, ListEnd,
-                ListItem, ListStart, MapEnd, MapItem, MapStart, OptionMarker, OuterSequenceEnd,
-                OuterSequenceItem, OuterSequenceStart, Panic, ProgramEnd, PushBool,
-                PushDate64FromNaiveStr, PushDate64FromUtcStr, PushDictionary, PushF16, PushF32,
-                PushF64, PushI16, PushI32, PushI64, PushI8, PushLargeUtf8, PushNull, PushU16,
-                PushU32, PushU64, PushU8, PushUtf8, TupleStructEnd, TupleStructItem,
-                TupleStructStart, UnionEnd, Variant,
-            },
+            bytecode::{dispatch_bytecode, Bytecode},
             compiler::{BufferCounts, Program, Structure},
         },
         sink::EventSink,
@@ -395,798 +386,6 @@ trait Instruction: std::fmt::Debug {
     }
 }
 
-impl Instruction for Panic {
-    const NAME: &'static str = "Panic";
-    const EXPECTED: &'static [&'static str] = &[];
-
-    fn accept_start_sequence(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_end_sequence(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_start_tuple(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_end_tuple(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_start_struct(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_end_struct(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_start_map(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_end_map(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_item(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_some(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_default(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_variant(
-        &self,
-        _: &Structure,
-        _: &mut MutableBuffers,
-        _: &str,
-        _: usize,
-    ) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_null(&self, _: &Structure, _: &mut MutableBuffers) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_bool(&self, _: &Structure, _: &mut MutableBuffers, _: bool) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_u8(&self, _: &Structure, _: &mut MutableBuffers, _: u8) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_u16(&self, _: &Structure, _: &mut MutableBuffers, _: u16) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_u32(&self, _: &Structure, _: &mut MutableBuffers, _: u32) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_u64(&self, _: &Structure, _: &mut MutableBuffers, _: u64) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_i8(&self, _: &Structure, _: &mut MutableBuffers, _: i8) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_i16(&self, _: &Structure, _: &mut MutableBuffers, _: i16) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_i32(&self, _: &Structure, _: &mut MutableBuffers, _: i32) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_i64(&self, _: &Structure, _: &mut MutableBuffers, _: i64) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_f32(&self, _: &Structure, _: &mut MutableBuffers, _: f32) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_f64(&self, _: &Structure, _: &mut MutableBuffers, _: f64) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-
-    fn accept_str(&self, _: &Structure, _: &mut MutableBuffers, _: &str) -> Result<usize> {
-        fail!("{}", self.message);
-    }
-}
-
-impl Instruction for MapEnd {
-    const NAME: &'static str = "MapEnd";
-    const EXPECTED: &'static [&'static str] = &["Item", "EndMap"];
-
-    fn accept_item(&self, structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].inc_current_items()?;
-        Ok(structure.maps[self.map_idx].key)
-    }
-
-    fn accept_end_map(&self, structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].push_current_items();
-        Ok(structure.maps[self.map_idx].r#return)
-    }
-}
-
-impl Instruction for OuterSequenceStart {
-    const NAME: &'static str = "OuterSequenceStart";
-    const EXPECTED: &'static [&'static str] = &["StartSequence", "StartTuple"];
-
-    fn accept_start_sequence(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-
-    fn accept_start_tuple(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for ProgramEnd {
-    const NAME: &'static str = "ProgramEnd";
-    const EXPECTED: &'static [&'static str] = &[];
-}
-
-impl Instruction for OuterSequenceItem {
-    const NAME: &'static str = "OuterSequenceItem";
-    const EXPECTED: &'static [&'static str] = &["EndSequence", "EndTuple", "Item"];
-
-    fn accept_end_sequence(
-        &self,
-        structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(structure.large_lists[self.list_idx].r#return)
-    }
-
-    fn accept_item(&self, _structure: &Structure, _buffers: &mut MutableBuffers) -> Result<usize> {
-        Ok(self.next)
-    }
-
-    fn accept_end_tuple(
-        &self,
-        structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(structure.large_lists[self.list_idx].r#return)
-    }
-}
-
-impl Instruction for OuterSequenceEnd {
-    const NAME: &'static str = "OuterSequenceEnd";
-    const EXPECTED: &'static [&'static str] = &["EndSequence", "EndTuple", "Item"];
-
-    fn accept_end_sequence(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-
-    fn accept_item(&self, structure: &Structure, _buffers: &mut MutableBuffers) -> Result<usize> {
-        Ok(structure.large_lists[self.list_idx].item)
-    }
-
-    fn accept_end_tuple(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for LargeListStart {
-    const NAME: &'static str = "LargeListStart";
-    const EXPECTED: &'static [&'static str] = &["StartSequence", "StartTuple"];
-
-    fn accept_start_sequence(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-
-    fn accept_start_tuple(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for LargeListItem {
-    const NAME: &'static str = "LargeListItem";
-    const EXPECTED: &'static [&'static str] = &["EndSequence", "EndTuple", "Item"];
-
-    fn accept_end_sequence(
-        &self,
-        structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u64_offsets[self.offsets].push_current_items();
-        Ok(structure.large_lists[self.list_idx].r#return)
-    }
-
-    fn accept_end_tuple(
-        &self,
-        structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u64_offsets[self.offsets].push_current_items();
-        Ok(structure.large_lists[self.list_idx].r#return)
-    }
-
-    fn accept_item(&self, _structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u64_offsets[self.offsets].inc_current_items()?;
-        Ok(self.next)
-    }
-}
-
-impl Instruction for LargeListEnd {
-    const NAME: &'static str = "LargeListEnd";
-    const EXPECTED: &'static [&'static str] = &["EndSequence", "EndTuple", "Item"];
-
-    fn accept_end_sequence(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u64_offsets[self.offsets].push_current_items();
-        Ok(self.next)
-    }
-
-    fn accept_end_tuple(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u64_offsets[self.offsets].push_current_items();
-        Ok(self.next)
-    }
-
-    fn accept_item(&self, structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u64_offsets[self.offsets].inc_current_items()?;
-        Ok(structure.large_lists[self.list_idx].item)
-    }
-}
-
-impl Instruction for ListStart {
-    const NAME: &'static str = "ListStart";
-    const EXPECTED: &'static [&'static str] = &["StartSequence", "StartTuple"];
-
-    fn accept_start_sequence(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-
-    fn accept_start_tuple(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for ListItem {
-    const NAME: &'static str = "ListItem";
-    const EXPECTED: &'static [&'static str] = &["EndSequence", "EndTuple", "Item"];
-
-    fn accept_end_sequence(
-        &self,
-        structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].push_current_items();
-        Ok(structure.lists[self.list_idx].r#return)
-    }
-
-    fn accept_end_tuple(
-        &self,
-        structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].push_current_items();
-        Ok(structure.lists[self.list_idx].r#return)
-    }
-
-    fn accept_item(&self, _structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].inc_current_items()?;
-        Ok(self.next)
-    }
-}
-
-impl Instruction for ListEnd {
-    const NAME: &'static str = "ListEnd";
-    const EXPECTED: &'static [&'static str] = &["EndSequence", "EndTuple", "Item"];
-
-    fn accept_end_sequence(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].push_current_items();
-        Ok(self.next)
-    }
-
-    fn accept_end_tuple(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].push_current_items();
-        Ok(self.next)
-    }
-
-    fn accept_item(&self, structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].inc_current_items()?;
-        Ok(structure.lists[self.list_idx].item)
-    }
-}
-
-impl Instruction for MapStart {
-    const NAME: &'static str = "MapStart";
-    const EXPECTED: &'static [&'static str] = &["StartMap"];
-
-    fn accept_start_map(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for MapItem {
-    const NAME: &'static str = "MapItem";
-    const EXPECTED: &'static [&'static str] = &["EndMap", "Item"];
-
-    fn accept_item(&self, _structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].inc_current_items()?;
-        Ok(self.next)
-    }
-
-    fn accept_end_map(&self, structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u32_offsets[self.offsets].push_current_items();
-        Ok(structure.maps[self.map_idx].r#return)
-    }
-}
-
-impl Instruction for TupleStructStart {
-    const NAME: &'static str = "TupleStructStart";
-    const EXPECTED: &'static [&'static str] = &["StartTuple"];
-
-    fn accept_start_tuple(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for TupleStructItem {
-    const NAME: &'static str = "TupleStructItem";
-    const EXPECTED: &'static [&'static str] = &["Item"];
-    fn accept_item(&self, _structure: &Structure, _buffers: &mut MutableBuffers) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-impl Instruction for TupleStructEnd {
-    const NAME: &'static str = "TupleStructEnd";
-    const EXPECTED: &'static [&'static str] = &["EndTuple"];
-
-    fn accept_end_tuple(
-        &self,
-        _structure: &Structure,
-        _buffers: &mut MutableBuffers,
-    ) -> Result<usize> {
-        Ok(self.next)
-    }
-}
-
-macro_rules! option_marker_handle {
-    ($name:ident$(, $($val:ident: $ty:ty),*)?) => {
-        fn $name(&self, structure: &Structure, buffers: &mut MutableBuffers $(, $($val: $ty),*)?) -> Result<usize> {
-            buffers.u1[self.validity].push(true);
-            dispatch_bytecode!(&structure.program[self.next], instr => instr.$name(structure, buffers $(, $($val),*)?))
-        }
-    };
-}
-
-/// Handle optionality markers (null / some)
-///
-/// The basic strategy is to keep this instruction active until any event but
-/// `Some` is encountered. If a `Null `event is encountered store a missing
-/// value and continue with the next field / item. If any other value is
-/// encountered, call the next instruction inline.
-///
-impl Instruction for OptionMarker {
-    const NAME: &'static str = "OptionMarker";
-    const EXPECTED: &'static [&'static str] = &[
-        "Some",
-        "Null",
-        "StartSequence",
-        "EndSequence",
-        "StartTuple",
-        "EndTuple",
-        "StartStruct",
-        "EndStruct",
-        "Item",
-        "Default",
-        "Variant",
-        "Bool",
-        "U8",
-        "U16",
-        "U32",
-        "U64",
-        "I8",
-        "I16",
-        "I32",
-        "I64",
-        "F32",
-        "F64",
-        "Str",
-    ];
-
-    fn accept_some(&self, _structure: &Structure, _buffers: &mut MutableBuffers) -> Result<usize> {
-        Ok(self.self_pos)
-    }
-
-    fn accept_null(&self, structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        apply_null(structure, buffers, self.null_definition)?;
-        Ok(self.if_none)
-    }
-
-    option_marker_handle!(accept_start_sequence);
-    option_marker_handle!(accept_end_sequence);
-    option_marker_handle!(accept_start_tuple);
-    option_marker_handle!(accept_end_tuple);
-    option_marker_handle!(accept_start_struct);
-    option_marker_handle!(accept_end_struct);
-    option_marker_handle!(accept_start_map);
-    option_marker_handle!(accept_end_map);
-    option_marker_handle!(accept_item);
-    option_marker_handle!(accept_default);
-    option_marker_handle!(accept_variant, name: &str, idx: usize);
-    option_marker_handle!(accept_bool, val: bool);
-    option_marker_handle!(accept_u8, val: u8);
-    option_marker_handle!(accept_u16, val: u16);
-    option_marker_handle!(accept_u32, val: u32);
-    option_marker_handle!(accept_u64, val: u64);
-    option_marker_handle!(accept_i8, val: i8);
-    option_marker_handle!(accept_i16, val: i16);
-    option_marker_handle!(accept_i32, val: i32);
-    option_marker_handle!(accept_i64, val: i64);
-    option_marker_handle!(accept_f32, val: f32);
-    option_marker_handle!(accept_f64, val: f64);
-    option_marker_handle!(accept_str, val: &str);
-}
-
-impl Instruction for Variant {
-    const NAME: &'static str = "Variant";
-    const EXPECTED: &'static [&'static str] = &["Variant"];
-
-    fn accept_variant(
-        &self,
-        structure: &Structure,
-        buffers: &mut MutableBuffers,
-        _name: &str,
-        idx: usize,
-    ) -> Result<usize> {
-        if idx < structure.unions[self.union_idx].fields.len() {
-            buffers.u8[self.type_idx].push(i8::try_from(idx)?.to_bytes());
-            Ok(structure.unions[self.union_idx].fields[idx])
-        } else {
-            fail!(
-                concat!(
-                    "Serialization failed: an unknown variant with index {child_idx} for field was ",
-                    "encountered. To fix this error, ensure all variants are seen during ",
-                    "schema tracing or add the relevant variants manually to the traced fields.",
-                ),
-                child_idx = idx,
-            )
-        }
-    }
-}
-
-impl Instruction for PushUtf8 {
-    const NAME: &'static str = "PushUtf8";
-    const EXPECTED: &'static [&'static str] = &["Str"];
-
-    fn accept_str(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-        val: &str,
-    ) -> Result<usize> {
-        buffers.u8[self.buffer].extend(val.as_bytes().iter().copied());
-        buffers.u32_offsets[self.offsets].push(val.len())?;
-        Ok(self.next)
-    }
-}
-
-impl Instruction for PushLargeUtf8 {
-    const NAME: &'static str = "PushLargeUtf8";
-    const EXPECTED: &'static [&'static str] = &["Str"];
-
-    fn accept_str(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-        val: &str,
-    ) -> Result<usize> {
-        buffers.u8[self.buffer].extend(val.as_bytes().iter().copied());
-        buffers.u64_offsets[self.offsets].push(val.len())?;
-        Ok(self.next)
-    }
-}
-
-impl Instruction for PushDate64FromNaiveStr {
-    const NAME: &'static str = "PushDate64FromNaiveStr";
-    const EXPECTED: &'static [&'static str] = &["Str"];
-
-    fn accept_str(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-        val: &str,
-    ) -> Result<usize> {
-        use chrono::NaiveDateTime;
-
-        buffers.u64[self.idx].push(val.parse::<NaiveDateTime>()?.timestamp_millis().to_bytes());
-        Ok(self.next)
-    }
-}
-
-impl Instruction for PushDate64FromUtcStr {
-    const NAME: &'static str = "PushDate64FromUtcStr";
-    const EXPECTED: &'static [&'static str] = &["Str"];
-
-    fn accept_str(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-        val: &str,
-    ) -> Result<usize> {
-        use chrono::{DateTime, Utc};
-
-        buffers.u64[self.idx].push(val.parse::<DateTime<Utc>>()?.timestamp_millis().to_bytes());
-        Ok(self.next)
-    }
-}
-
-impl Instruction for PushDictionary {
-    const NAME: &'static str = "PushDictionary";
-    const EXPECTED: &'static [&'static str] = &["Str"];
-
-    fn accept_str(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-        val: &str,
-    ) -> Result<usize> {
-        use {DictionaryIndex as I, DictionaryValue as V};
-
-        let idx = if buffers.dictionaries[self.dictionary].contains_key(val) {
-            buffers.dictionaries[self.dictionary][val]
-        } else {
-            match self.values {
-                V::Utf8 { buffer, offsets } => {
-                    buffers.u8[buffer].extend(val.as_bytes().iter().copied());
-                    buffers.u32_offsets[offsets].push(val.len())?;
-                }
-                V::LargeUtf8 { buffer, offsets } => {
-                    buffers.u8[buffer].extend(val.as_bytes().iter().copied());
-                    buffers.u64_offsets[offsets].push(val.len())?;
-                }
-            }
-
-            let idx = buffers.dictionaries[self.dictionary].len();
-            buffers.dictionaries[self.dictionary].insert(val.to_string(), idx);
-            idx
-        };
-
-        match self.indices {
-            I::U8(indices) => buffers.u8[indices].push(idx.try_into()?),
-            I::U16(indices) => buffers.u16[indices].push(idx.try_into()?),
-            I::U32(indices) => buffers.u32[indices].push(idx.try_into()?),
-            I::U64(indices) => buffers.u64[indices].push(idx.try_into()?),
-            I::I8(indices) => buffers.u8[indices].push(i8::try_from(idx)?.to_bytes()),
-            I::I16(indices) => buffers.u16[indices].push(u16::try_from(idx)?.to_bytes()),
-            I::I32(indices) => buffers.u32[indices].push(u32::try_from(idx)?.to_bytes()),
-            I::I64(indices) => buffers.u64[indices].push(u64::try_from(idx)?.to_bytes()),
-        }
-        Ok(self.next)
-    }
-}
-
-impl Instruction for UnionEnd {
-    const NAME: &'static str = "UnionEnd";
-    const EXPECTED: &'static [&'static str] = &[];
-}
-
-impl Instruction for PushNull {
-    const NAME: &'static str = "PushNull";
-    const EXPECTED: &'static [&'static str] = &["Null"];
-
-    fn accept_null(&self, _structure: &Structure, buffers: &mut MutableBuffers) -> Result<usize> {
-        buffers.u0[self.idx].push(());
-        Ok(self.next)
-    }
-}
-
-macro_rules! impl_primitive_instruction {
-    (
-        $(
-            $name:ident($val_type:ty, $builder:ident) {
-                $($func:ident($ty:ty),)*
-            },
-        )*
-    ) => {
-        $(
-            impl Instruction for $name {
-                const NAME: &'static str = stringify!($name);
-                const EXPECTED: &'static [&'static str] = &[$(stringify!($ty)),*];
-
-                $(
-                    fn $func(&self, _structure: &Structure, buffers: &mut MutableBuffers, val: $ty) -> Result<usize> {
-                        let val = <$val_type>::try_from(val)?;
-                        buffers.$builder[self.idx].push(ToBytes::to_bytes(val));
-                        Ok(self.next)
-                    }
-                )*
-            }
-        )*
-    };
-}
-
-impl_primitive_instruction!(
-    PushU8(u8, u8) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushU16(u16, u16) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushU32(u32, u32) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushU64(u64, u64) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushI8(i8, u8) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushI16(i16, u16) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushI32(i32, u32) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushI64(i64, u64) {
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-    PushF16(WrappedF16, u16) {
-        accept_f32(f32),
-        accept_f64(f64),
-    },
-    PushF32(WrappedF32, u32) {
-        accept_f32(f32),
-        accept_f64(f64),
-    },
-    PushF64(WrappedF64, u64) {
-        accept_f32(f32),
-        accept_f64(f64),
-        accept_u8(u8),
-        accept_u16(u16),
-        accept_u32(u32),
-        accept_u64(u64),
-        accept_i8(i8),
-        accept_i16(i16),
-        accept_i32(i32),
-        accept_i64(i64),
-    },
-);
-
-impl Instruction for PushBool {
-    const NAME: &'static str = "PushBool";
-    const EXPECTED: &'static [&'static str] = &["Bool"];
-
-    fn accept_bool(
-        &self,
-        _structure: &Structure,
-        buffers: &mut MutableBuffers,
-        val: bool,
-    ) -> Result<usize> {
-        buffers.u1[self.idx].push(val);
-        Ok(self.next)
-    }
-}
-
 macro_rules! dispatch_instruction {
     ($this:expr, $method:ident) => {
         {
@@ -1356,34 +555,4 @@ impl EventSink for Interpreter {
 
         Ok(())
     }
-}
-
-macro_rules! apply_null {
-    ($structure:expr, $buffers:expr, $null_definition:expr, $name:ident) => {
-        for &idx in &$structure.nulls[$null_definition].$name {
-            $buffers.$name[idx].push(Default::default());
-        }
-    };
-}
-
-fn apply_null(
-    structure: &Structure,
-    buffers: &mut MutableBuffers,
-    null_definition: usize,
-) -> Result<()> {
-    apply_null!(structure, buffers, null_definition, u0);
-    apply_null!(structure, buffers, null_definition, u1);
-    apply_null!(structure, buffers, null_definition, u8);
-    apply_null!(structure, buffers, null_definition, u16);
-    apply_null!(structure, buffers, null_definition, u32);
-    apply_null!(structure, buffers, null_definition, u64);
-
-    for &idx in &structure.nulls[null_definition].u32_offsets {
-        buffers.u32_offsets[idx].push_current_items();
-    }
-    for &idx in &structure.nulls[null_definition].u64_offsets {
-        buffers.u64_offsets[idx].push_current_items();
-    }
-
-    Ok(())
 }
